Track Names
There are 200000 in the big one, but some of them have non-utf8 encoded characters that python can't handle in windows (like a Chinese/Korean character) so I included the 55,000 track long file that has only windows readable characters. To be honest if we want to limit ourselves to english only songs, that might be the one to go with.